requirements.txt
streamlit
langchain
langchain-pinecone
sentence-transformers
tiktoken
octoai
octoai-sdk
python-dotenv
playwright
bs4
pinecone